+++ src/test/java/com/epam/reportportal/cucumber/FailedTest.java
@@ -25,6 +25,7 @@ import com.epam.reportportal.listeners.ListenerParameters;
 import com.epam.reportportal.service.ReportPortal;
 import com.epam.reportportal.service.ReportPortalClient;
 import com.epam.reportportal.util.test.CommonUtils;
+import com.epam.reportportal.utils.formatting.MarkdownUtils;
 import com.epam.ta.reportportal.ws.model.FinishTestItemRQ;
 import com.epam.ta.reportportal.ws.model.log.SaveLogRQ;
 import cucumber.api.CucumberOptions;
@@ -39,12 +40,11 @@ import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.stream.Collectors;
 
 import static com.epam.reportportal.cucumber.integration.util.TestUtils.filterLogs;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.hasSize;
+import static org.hamcrest.Matchers.*;
+import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.*;
 
 /**
@@ -53,19 +53,25 @@ import static org.mockito.Mockito.*;
 public class FailedTest {
 
 	private static final String EXPECTED_ERROR = "java.lang.IllegalStateException: " + FailedSteps.ERROR_MESSAGE;
+	private static final String EXPECTED_STACK_TRACE = EXPECTED_ERROR
+			+ "\n\tat com.epam.reportportal.cucumber.integration.feature.FailedSteps.i_have_a_failed_step(FailedSteps.java:31)"
+			+ "\n\tat ✽.I have a failed step(src/test/resources/features/FailedScenario.feature:4)\n";
+	private static final String ERROR_LOG_TEXT = "Error:\n" + EXPECTED_STACK_TRACE;
+
+	private static final String SCENARIO_CODE_REFERENCES_WITH_ERROR = MarkdownUtils.asTwoParts("src/test/resources/features/FailedScenario.feature",
+			ERROR_LOG_TEXT
+	);
 
 	@CucumberOptions(features = "src/test/resources/features/FailedScenario.feature", glue = {
 			"com.epam.reportportal.cucumber.integration.feature" }, plugin = { "pretty",
 			"com.epam.reportportal.cucumber.integration.TestScenarioReporter" })
-	public static class FailedScenarioReporter extends AbstractTestNGCucumberTests {
-
+	public static class FailedScenarioReporterTest extends AbstractTestNGCucumberTests {
 	}
 
 	@CucumberOptions(features = "src/test/resources/features/FailedScenario.feature", glue = {
 			"com.epam.reportportal.cucumber.integration.feature" }, plugin = { "pretty",
 			"com.epam.reportportal.cucumber.integration.TestStepReporter" })
-	public static class FailedStepReporter extends AbstractTestNGCucumberTests {
-
+	public static class FailedStepReporterTest extends AbstractTestNGCucumberTests {
 	}
 
 	private final String launchId = CommonUtils.namedId("launch_");
@@ -73,7 +79,7 @@ public class FailedTest {
 	private final String testId = CommonUtils.namedId("test_");
 	private final String stepId = CommonUtils.namedId("step_");
 	private final String nestedStepId = CommonUtils.namedId("nested_");
-	private final List<Pair<String,String>> nestedSteps = Collections.singletonList(Pair.of(stepId, nestedStepId));
+	private final List<Pair<String, String>> nestedSteps = Collections.singletonList(Pair.of(stepId, nestedStepId));
 
 	private final ListenerParameters parameters = TestUtils.standardParameters();
 	private final ReportPortalClient client = mock(ReportPortalClient.class);
@@ -92,7 +98,7 @@ public class FailedTest {
 	@Test
 	@SuppressWarnings("unchecked")
 	public void verify_failed_step_reporting_scenario_reporter() {
-		TestUtils.runTests(FailedScenarioReporter.class);
+		TestUtils.runTests(FailedScenarioReporterTest.class);
 
 		verify(client).startTestItem(any());
 		verify(client).startTestItem(same(suiteId), any());
@@ -104,7 +110,7 @@ public class FailedTest {
 		verify(client).finishTestItem(same(testId), finishCaptor.capture());
 
 		List<FinishTestItemRQ> finishRqs = finishCaptor.getAllValues();
-		finishRqs.subList(0, finishRqs.size() - 1).forEach(e->assertThat(e.getStatus(), equalTo(ItemStatus.FAILED.name())));
+		finishRqs.subList(0, finishRqs.size() - 1).forEach(e -> assertThat(e.getStatus(), equalTo(ItemStatus.FAILED.name())));
 
 		ArgumentCaptor<List<MultipartBody.Part>> logCaptor = ArgumentCaptor.forClass(List.class);
 		verify(client, atLeastOnce()).log(logCaptor.capture());
@@ -118,7 +124,7 @@ public class FailedTest {
 	@Test
 	@SuppressWarnings("unchecked")
 	public void verify_failed_step_reporting_step_reporter() {
-		TestUtils.runTests(FailedStepReporter.class);
+		TestUtils.runTests(FailedStepReporterTest.class);
 
 		verify(client).startTestItem(any());
 		verify(client).startTestItem(same(suiteId), any());
@@ -127,7 +133,7 @@ public class FailedTest {
 		verify(client).finishTestItem(same(stepId), finishCaptor.capture());
 		verify(client).finishTestItem(same(testId), finishCaptor.capture());
 
-		finishCaptor.getAllValues().forEach(e->assertThat(e.getStatus(), equalTo(ItemStatus.FAILED.name())));
+		finishCaptor.getAllValues().forEach(e -> assertThat(e.getStatus(), equalTo(ItemStatus.FAILED.name())));
 
 		ArgumentCaptor<List<MultipartBody.Part>> logCaptor = ArgumentCaptor.forClass(List.class);
 		verify(client, atLeastOnce()).log(logCaptor.capture());
@@ -135,5 +141,44 @@ public class FailedTest {
 		List<SaveLogRQ> expectedErrorList = filterLogs(logCaptor, l -> l.getMessage() != null && l.getMessage().startsWith(EXPECTED_ERROR));
 		SaveLogRQ expectedError = expectedErrorList.get(0);
 		assertThat(expectedError.getItemUuid(), equalTo(stepId));
+	}
+
+	@Test
+	public void verify_failed_nested_step_description_scenario_reporter() {
+		TestUtils.runTests(FailedScenarioReporterTest.class);
+
+		verify(client).startTestItem(any());
+		verify(client).startTestItem(same(suiteId), any());
+		verify(client).startTestItem(same(testId), any());
+		verify(client).startTestItem(same(stepId), any());
+		ArgumentCaptor<FinishTestItemRQ> finishCaptor = ArgumentCaptor.forClass(FinishTestItemRQ.class);
+		verify(client).finishTestItem(same(nestedStepId), finishCaptor.capture());
+		verify(client).finishTestItem(same(testId), finishCaptor.capture());
+
+		List<FinishTestItemRQ> finishRqs = finishCaptor.getAllValues();
+		finishRqs.subList(0, finishRqs.size() - 1).forEach(e -> assertThat(e.getStatus(), equalTo(ItemStatus.FAILED.name())));
+
+		FinishTestItemRQ step = finishRqs.get(0);
+		assertThat(step.getDescription(), not(equalTo(ERROR_LOG_TEXT)));
+	}
+
+	@Test
+	public void verify_failed_step_description_step_reporter() {
+		TestUtils.runTests(FailedStepReporterTest.class);
+
+		verify(client).startTestItem(any());
+		verify(client).startTestItem(same(suiteId), any());
+		verify(client).startTestItem(same(testId), any());
+		ArgumentCaptor<FinishTestItemRQ> finishCaptor = ArgumentCaptor.forClass(FinishTestItemRQ.class);
+		verify(client).finishTestItem(same(stepId), finishCaptor.capture());
+		verify(client).finishTestItem(same(testId), finishCaptor.capture());
+
+		List<FinishTestItemRQ> finishRqs = finishCaptor.getAllValues();
+		finishRqs.forEach(e -> assertThat(e.getStatus(), equalTo(ItemStatus.FAILED.name())));
+
+		FinishTestItemRQ step = finishRqs.get(0);
+		assertThat(step.getDescription(), equalTo(ERROR_LOG_TEXT));
+		FinishTestItemRQ test = finishRqs.get(1);
+		assertThat(test.getDescription(), equalTo(SCENARIO_CODE_REFERENCES_WITH_ERROR));
 	}
 }