+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Common Stack Trace frames skip in description and logs, by @HardNorth
+- Reporting of Last Error Log in Item description, by @HardNorth and @ArtemOAS
+### Changed
+- Client version updated on [5.2.21](https://github.com/reportportal/client-java/releases/tag/5.2.21), by @HardNorth
 
 ## [5.2.2]
 ### Changed
 - Client version updated on [5.2.13](https://github.com/reportportal/client-java/releases/tag/5.2.13), by @HardNorth
 ### Removed
 - `OkHttp` dependency, by @HardNorth
+- JSR-305 dependency, by @HardNorth
 
 ## [5.2.1]
 ### Changed